Create K8version-ami-check
This automation retrieves the ami-ids of the worker nodes and compare them with the latest recommended ami- ids corresponding to user's Cluster version and region. If there is a mismatch, it recommends the user the latest ami-id to be used. The automation also caters to ami with gpu support.
It will work as one of the best practices to be recommended to the users facing : 'node group switching between ready and not ready state' error.